PERF: use pyarrow kernel for Series.str.normalize with ArrowStringArray
ArrowStringArray lists ObjectStringArrayMixin ahead of ArrowExtensionArray
in its MRO, so any _str_* method missing from the alias block silently gets
the object-dtype implementation. _str_normalize had an Arrow implementation
that was unreachable for StringDtype with storage "pyarrow".

Move _str_normalize to ArrowStringArrayMixin, next to _str_pad, and alias it
from there so it matches every neighbouring alias in the block. The mixin
already declares the _apply_elementwise stub, so the fallback path carries
over unchanged; ArrowExtensionArray inherits the mixin and keeps the method.

This is a performance-only change: StringDtype("pyarrow") was already
correct, because the object path calls unicodedata.normalize. At 500k
elements the decomposing forms speed up between 1.5x and 17x depending on
payload -- 60.1ms -> 3.7ms for pure ASCII, 124.8ms -> 31.6ms for short mixed
strings, 197.2ms -> 118.6ms for longer mixed strings. The win is largest
when the kernel can short-circuit and smallest when the data genuinely needs
decomposing. It also closes the gap with ArrowDtype, which reaches the same
code today.

The composing forms NFC/NFKC keep the unicodedata fallback and are
unchanged: pc.utf8_normalize skips the canonical composition step and
silently returns decomposed output (still reproducible on pyarrow 23.0.1;
worked around for ArrowDtype in #66221).

Validate the normalization form so an invalid one keeps raising
ValueError("invalid normalization form") instead of pyarrow's message once
the Arrow path is reachable. This also aligns ArrowDtype, which raises
pyarrow's message today.
